Use type literal for path
Using string literal type _should_ not break anything (tested on a
larger project) since it is a subtype of string.

Using literal instead of generic string allows to do type-safe
RPC matching, for example it generic middlewares for cache, ACL etc.
Without it we need to match a string for path (long, prone to errors) or
method names (ambiguous) and have no way to check, wheather the matching
is correct or not.

With literals, we can extract all paths from generated stubs with
meta-typing and use it in these occasions.